refactor(oscript): вынести детект &Обходимое в отдельный OScriptIterable
&Обходимое — маркер стандартной библиотеки самого OneScript (движковый
UserIterableContextInstance), к библиотеке extends отношения не имеет.
Перенёс isIterable из OScriptExtends в новый компонент OScriptIterable
(пакет types.oscript), самодостаточный (сканирует аннотации конструктора
через Methods.isOscriptClassConstructorName). OScriptModuleMembersProvider
теперь зависит от OScriptIterable.

Тесты isIterable переехали в OScriptIterableTest.
